requirements.txt
ststreamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
ortools>=9.7.0